refactor(plugins): split runtime deps materialization
